Fix workflow input deserializer
