+++ lib/cadence/version.rb
@@ -1,3 +1,3 @@
 module Cadence
-  VERSION = '0.1.16'.freeze
+  VERSION = '0.1.17'.freeze
 end
